+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Removed
-- Compatibility with TYPO3 v11
+- Compatibility with TYPO3 v11 (#16)
+- Compatibility with TYPO3 v12 (#28)
 
 ## [2.4.0] - 2025-09-17